[fix] conflict 해결을 위한 TraiinerController 제작
